makefile fixed for docker
